\name{lintul0}

\alias{lintul0}

\title{
LINTUL0 crop growth model
}

\description{
LINTUL is a relatively simple crop growth simulation model
}

\usage{
lintul0(crop, weather, control)
lintul0_crop()
}

\arguments{
\item{crop}{list. Crop parameters}
\item{weather}{data.frame with weather data}
\item{control}{list. Model control options}
}

\value{
matrix
}

\references{
Spitters, C.J.T. and A.H.C.M. Schapendonk, 1990. Evaluation of breeding strategies for drought tolerance in potato by means of crop growth simulation. Plant and Soil 123: 193-203.
}


