\name{LINTUL-class}

\docType{class}


\alias{DailyWeather}
\alias{Rcpp_DailyWeather}

\alias{Lintul1Control}
\alias{Lintul1Crop}
\alias{Lintul1Model}
\alias{Lintul1Output}
\alias{Rcpp_Lintul1Crop}
\alias{Rcpp_Lintul1Control}
\alias{Rcpp_Lintul1Model}
\alias{Rcpp_Lintul1Output}

\alias{Lintul2Control}
\alias{Lintul2Crop}
\alias{Lintul2Soil}
\alias{Lintul2Model}
\alias{Lintul2Output}
\alias{Rcpp_Lintul2Crop}
\alias{Rcpp_Lintul2Soil}
\alias{Rcpp_Lintul2Control}
\alias{Rcpp_Lintul2Output}
\alias{Rcpp_Lintul2Model}

\alias{Lintul3Crop}
\alias{Lintul3Soil}
\alias{Lintul3Control}
\alias{Lintul3Model}
\alias{Lintul3Output}
\alias{Rcpp_Lintul3Crop}
\alias{Rcpp_Lintul3Soil}
\alias{Rcpp_Lintul3Control}
\alias{Rcpp_Lintul3Output}
\alias{Rcpp_Lintul3Model}


\alias{DailyWeather-class}
\alias{Rcpp_DailyWeather-class}

\alias{Lintul1Control-class}
\alias{Lintul1Crop-class}
\alias{Lintul1Model-class}
\alias{Lintul1Output-class}
\alias{Rcpp_Lintul1Crop-class}
\alias{Rcpp_Lintul1Control-class}
\alias{Rcpp_LintulWeather-class}
\alias{Rcpp_Lintul1Model-class}
\alias{Rcpp_Lintul1Output-class}

\alias{Lintul2Control-class}
\alias{Lintul2Crop-class}
\alias{Lintul2Soil-class}
\alias{Lintul2Model-class}
\alias{Lintul2Output-class}
\alias{Rcpp_Lintul2Crop-class}
\alias{Rcpp_Lintul2Soil-class}
\alias{Rcpp_Lintul2Control-class}
\alias{Rcpp_Lintul2Output-class}
\alias{Rcpp_Lintul2Model-class}

\alias{Lintul3Crop-class}
\alias{Lintul3Soil-class}
\alias{Lintul3Control-class}
\alias{Lintul3Model-class}
\alias{Lintul3Output-class}
\alias{Rcpp_Lintul3Crop-class}
\alias{Rcpp_Lintul3Soil-class}
\alias{Rcpp_Lintul3Control-class}
\alias{Rcpp_Lintul3Output-class}
\alias{Rcpp_Lintul3Model-class}

\title{LINTUL model classes}

\description{
These are the classes!
}

\section{Objects from the Class}{
Objects can be created with these functions: \code{\link{lintul1}}, \code{\link{lintul2}}, \code{\link{lintul3}}.
}

\examples{
x <- lintul1()
}

\keyword{classes}

