\name{LINTUL-classes}

\docType{class}

\alias{LintulControl}
\alias{LintulWeather}

\alias{Lintul1Crop}
\alias{Lintul1Model}
\alias{Lintul1Output}
\alias{Rcpp_Lintul1Crop}
\alias{Rcpp_LintulControl}
\alias{Rcpp_LintulWeather}
\alias{Rcpp_Lintul1Model}
\alias{Rcpp_Lintul1Output}

\alias{Lintul2Crop}
\alias{Lintul2Soil}
\alias{Lintul2Model}
\alias{Lintul2Output}
\alias{Rcpp_Lintul2Crop}
\alias{Rcpp_Lintul2Soil}
\alias{Rcpp_Lintu2Control}
\alias{Rcpp_Lintul2Output}
\alias{Rcpp_Lintul2Model}


\title{LINTUL model classes}

\description{
These are the classes!
}


\keyword{classes}
